feat(editor): bind scenes to subgraphs + unified View picker
A scene is the physical realization of one hierarchy level. A floor
plan of "Building A / Floor 1" should only show the gear that lives
in that subgraph — not VPCs, not nodes from other floors. Without a
scope, every scene rendered the entire NetworkGraph and the user had
to manually hide everything irrelevant.

Data model
- Scene gains optional scopeSubgraphId. undefined = root scope (whole
  graph, fine for projects with no enclosing subgraph). A subgraph id
  filters visible nodes/links to descendants of that subgraph (via
  Node.parent / Subgraph.parent transitive closure).
- Existing scenes without scopeSubgraphId stay root-scoped — no
  migration needed, behavior unchanged.

UI consolidation
- SheetBar (hierarchy drill) and SceneBar (scene picker) both lived
  bottom-center and overlapped semantically once scenes carry scope.
  Merged into a single ViewPicker that lays out:
    Root
      Diagram (logical)
      <root-scoped scenes>
      + Add scene at root…
    Subgraphs
      Building A
        <its scenes>
        + Add scene…
      Cloud  (logical — no scene action)
- isPhysicalSubgraph(sg) gates the "Add scene" affordance: hardware
  or unspecified subgraphs allow scenes; compute/service do not.
- SheetBar.svelte and SceneBar.svelte are deleted.

Behavior
- Switching to a scene whose scope still references a removed
  subgraph degrades to "no nodes visible" — preferable to silently
  reverting to root, since the user explicitly bound it.
- Calibration / wire authoring / pin rendering unchanged.
